+++ app/Data/MessageData.php
@@ -7,6 +7,9 @@ class MessageData
     const TYPE_TEXT = 'text';
     const TYPE_STRUCTURE = 'structure';
 
+    const SCOPE_GROUP = 'group';
+    const SCOPE_PRIVATE = 'private';
+
     const CATEGORY_ID_TEXT = 1;
     const CATEGORY_ID_ACCEPTED_BOOKING_TIME = 2;
     const CATEGORY_ID_ACCEPTED_PACKAGE_BOOKING = 3;