+++ app/Entities/GroupGlobalSetting.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Entities;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class GroupGlobalSetting extends Model
+{
+    use HasFactory;
+}